+++ src/main/java/com/brand/blockus/registry/content/bundles/DyedBSSWBundle.java
@@ -2,6 +2,7 @@ package com.brand.blockus.registry.content.bundles;
 
 import com.brand.blockus.utils.blocks.ColorBlockItemCollection;
 import com.brand.blockus.utils.helper.BlockFactory;
+import net.minecraft.references.BlockItemId;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockBehaviour;
@@ -15,7 +16,11 @@ public record DyedBSSWBundle(
     ColorBlockItemCollection block,
     ColorBlockItemCollection stairs,
     ColorBlockItemCollection slab,
-    ColorBlockItemCollection wall
+    ColorBlockItemCollection wall,
+    ColorCollection<BlockItemId> blockId,
+    ColorCollection<BlockItemId> stairsId,
+    ColorCollection<BlockItemId> slabId,
+    ColorCollection<BlockItemId> wallId
 ) {
 
     public static final List<DyedBSSWBundle> LIST = new ArrayList<>();
@@ -37,10 +42,12 @@ public record DyedBSSWBundle(
     public static <Id> DyedBSSWBundle register(ColorCollection<Id> ids, ColorCollection<Id> idsStairs, ColorCollection<Id> idsSlab, ColorCollection<Id> idsWall, TriFunction<Id, Function<BlockBehaviour.Properties, Block>, BlockBehaviour.Properties, Block> register, Provider provider, boolean includeWall) {
         ColorBlockItemCollection block = BlockFactory.registerDyedBlocks(ids, register, provider::properties);
         DyedBSSWBundle bundle = new DyedBSSWBundle(
+
             block,
             BlockFactory.registerDyedBlocks(idsStairs, register, (color, p) -> new StairBlock(block.blocks().pick(color).defaultBlockState(), p), BlockFactory.copyDyedBlocks(block.blocks())),
             BlockFactory.registerDyedBlocks(idsSlab, register, (var0, p) -> new SlabBlock(p), BlockFactory.copyDyedBlocks(block.blocks())),
-            includeWall ? BlockFactory.registerDyedBlocks(idsWall, register, (var0, p) -> new WallBlock(p), BlockFactory.copyDyedBlocks(block.blocks())) : null
+            includeWall ? BlockFactory.registerDyedBlocks(idsWall, register, (var0, p) -> new WallBlock(p), BlockFactory.copyDyedBlocks(block.blocks())) : null,
+            (ColorCollection<BlockItemId>) ids, (ColorCollection<BlockItemId>) idsStairs, (ColorCollection<BlockItemId>) idsSlab, (ColorCollection<BlockItemId>) idsWall
         );
         LIST.add(bundle);
         return bundle;