+++ changelog.md
@@ -1,4 +1,5 @@
 * Fixed gingham wool missing in the creative inventory
+* Fixed missing asphalt stairs/slab, netherite stairs/slab, cut soul sandstone slab, smooth stone stairs & ice brick wall in stairs, slabs & wall tags 
 * Added `#blockus:lit_redstone_lamps` block/item tags
 * Updated zh translations by SolidBlocks